+++ tests/testthat/test-vocab-high-level.R
@@ -23,11 +23,11 @@ test_that("Vocabulary with foreach", {
   iterator_list = mapply(function(x, i) get_test_iterator(x, i), txt_splits, ids_splits)
   expect_warning(vocab_2 <- create_vocabulary(iterator_list))
 
-  expect_equal(sort(vocab_1$vocab$terms),  sort(vocab_2$vocab$terms))
-  expect_equal(vocab_1$document_count,  vocab_2$document_count)
-  expect_equal(vocab_1$document_count,  max(train_ind))
-  expect_equal(vocab_1$ngram,  vocab_2$ngram)
+  expect_equal(sort(vocab_1$terms),  sort(vocab_2$terms))
+  expect_equal(attr(vocab_1, "document_count"),  attr(vocab_2, "document_count"))
+  expect_equal(attr(vocab_1, "document_count"),  max(train_ind))
+  expect_equal(attr(vocab_1, "ngram"), attr(vocab_2, "ngram"))
 
-  expect_equal(create_dtm(iterator, vocab_vectorizer(vocab_1)),
-               create_dtm(iterator_list, vocab_vectorizer(vocab_1)))
+  expect_warning(dtm11 <- create_dtm(iterator_list, vocab_vectorizer(vocab_1)))
+  expect_equal(create_dtm(iterator, vocab_vectorizer(vocab_1)), dtm11)
 })